[kitsune] Create a Kitsune-specific mechanism to register runtime functions
These originally registered themselves as LLVM's runtime functions, but they
are now completely independent of this mechanism. This should make it easier to
add register new runtime functions.

The lowerig of the kitsune mobile alloc and free intrinsics for the non-GPU
tapir targets were changed to use the runtime's memory allocation function. This
registers the allocation with Kitsune's runtime. It is not clear that there is
any advantage to doing this, but it may be useful in the future when we
to multi-target allocation.

+++ kitsune/runtime/memory.cpp
@@ -74,7 +74,7 @@ template <> constexpr const char *getTypeFmt<int64_t>() { return "%ld"; }
 template <> constexpr const char *getTypeFmt<float>() { return "%f"; }
 template <> constexpr const char *getTypeFmt<double>() { return "%g"; }
 
-extern "C" [[gnu::malloc]] void *__kitrt_default_mem_alloc(size_t bytes) {
+extern "C" [[gnu::malloc]] void *__kitrt_default_mem_alloc(uint64_t bytes) {
   void *ptr = malloc(bytes);
   __kitrt_register_mem_alloc(ptr, bytes);
   return ptr;
